Separate RV, SPI, & UART clocks
Also add a first draft of the clocking constraints
for DesignCompiler (constrs/riscvcore_silicon.sdc)

Reworked the clock divider a bit to make it more
straight-forward. Instead of giving it a constant
as a signal, it now takes a divisor (must be an
even number) as a parameter. This allows it to
mirror the divisors in the new constraints file.

The biggest problem before was that both the UART
and SPI controllers were using the 50MHz clock
instead of the 12.5MHz and 115200 kHz clocks as
they should. I believe this fixes that.