Fix errors causing nonfunctional regfile and 7seg display
Changes to Decode.sv in commit "Fix synthesis errors with DC, ready for trial" causes core to be unable to write to registers. This fixes that. This also fixes an error in printing to 7seg display, but 7seg display still requires further fixes, most likely from issues arising from the new clock dividers.